Added simple mutators. Added RNG interface.
Modified mmaped IO so that only 4Gb and less files are used. Won't fuzz
that big files anyway.

+++ src/Fuzzman.Core/System/Mmap/MappedFile.cs
@@ -28,7 +28,7 @@ namespace Fuzzman.Core.System.Mmap
             }
         }
 
-        public MappedFileView CreateView(ulong offset, ulong length)
+        public MappedFileView CreateView(uint offset, uint length)
         {
             return new MappedFileView(this.mappingHandle, this.mappingAccess, offset, length);
         }
